+++ src/utils/typed_array.hpp
@@ -46,6 +46,7 @@ namespace node_gdal {
 namespace TypedArray {
 
 Local<Value> New(GDALDataType type, unsigned int length);
+Local<Value> New(GDALDataType type, void *data, unsigned int length);
 GDALDataType Identify(Local<Object> array);
 void *Validate(Local<Object> obj, GDALDataType type, int min_length);
 bool ValidateLength(int length, int min_length);